+++ app.py
@@ -1,7 +1,7 @@
+from __version__ import __version__
 import json
 import LDMigrate
 import LDConfig
-
 
 config = LDConfig.LDConfig("app.ini")
 settings = config.get_config()
@@ -10,9 +10,11 @@ ldmigrator = LDMigrate.LDMigrate(
     settings["source_api_token"],
     settings["source_project_key"],
     settings["target_api_token"],
-    settings["target_project_key"],
+    project_key_target=settings["target_project_key"],
+    flags_to_ignore=settings["flags_to_ignore"],
+    migration_mode=settings["migration_mode"]
 )
 
 result = ldmigrator.migrate()
 
-print(result)
+print(json.dumps(result))